20 x 20 and format up date
the 20x20 grid i feel like my  logic is correct with defining each way to get a product my limiting factor is knoledgebase i dont know how to define what i want. when its going though the array grid and it gets to a diagonal that goes out of the grid it brakes im not sure how to say when you get out of the grid youre done.