JS object programing
i think the reason it keeps returning an scientific notation is because JS is an Object oranted prograimming language. i think what i was doing was more funtional. i think you could do this in JS but im not sure how to get JS to ack like a funtional programming language